pyproject.toml
[project]
name = "stride"
version = "0.1.0"
description = "STRIDE: Stride-level Tracking to Reveal Impairment before Dementia Emerges"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Yipeng Li", email = "y9li@ucsd.edu"},
]
dependencies = [
    "click>=8.0",
    "numpy>=1.24",
    "opencv-python>=4.8",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "seaborn>=0.13",
    "shapely>=2.0",
    "sleap-io>=0.5.4",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
]

[project.scripts]
stride = "stride.cli:cli"

[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]
